# internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the UnderNET Channel Service. It provides a RESTful interface for managing users, channels, and other resources. \u003c!-- ReDoc-Inject: \u003csecurity-definitions\u003e --\u003e\n# Authorization\n**JWT Bearer Token:** The main authorization method for the API. Needs `Authorization: Bearer \u003cjwt-token\u003e` HTTP header to authenticate.\n\u003c!-- ReDoc-Inject: \u003csecurity-definitions\u003e --\u003e",
        "title": "UnderNET Channel Service API",
        "contact": {
            "name": "Ratler",
            "email": "ratler@undernet.org"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/UndernetIRC/cservice-api/blob/master/LICENSE"
        },
        "version": "0.1"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Get current user information with detailed channel membership data\nPerformance: Uses optimized single-query approach to avoid N+1 problems",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get current user information",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserResponse"
                        }
                    },
                    "401": {
                        "description": "Authorization information is missing or invalid."
                    },
                    "404": {
                        "description": "User not found."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/user/2fa/activate": {
            "post": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Validates the OTP code and activates 2FA for the user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Complete 2FA enrollment",
                "parameters": [
                    {
                        "description": "OTP code for 2FA activation",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ActivateTOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "2FA activated successfully"
                    },
                    "400": {
                        "description": "Bad request - validation failed"
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token"
                    },
                    "403": {
                        "description": "Forbidden - invalid OTP code"
                    },
                    "409": {
                        "description": "Conflict - 2FA already enabled or not enrolled"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/2fa/disable": {
            "post": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Disables 2FA for the user account. Requires both current password and valid OTP code for security.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Disable 2FA",
                "parameters": [
                    {
                        "description": "Password and OTP code for 2FA disabling",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.DisableTOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "2FA disabled successfully"
                    },
                    "400": {
                        "description": "Bad request - validation failed"
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token"
                    },
                    "403": {
                        "description": "Forbidden - incorrect password or invalid OTP"
                    },
                    "409": {
                        "description": "Conflict - 2FA is not enabled"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/2fa/enroll": {
            "post": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Generates a QR code and secret for TOTP 2FA enrollment. Requires current password for security.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Start 2FA enrollment",
                "parameters": [
                    {
                        "description": "Password confirmation for 2FA enrollment",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.EnrollTOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.EnrollTOTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation failed"
                    },
                    "401": {
                        "description": "Unauthorized - missing or invalid token"
                    },
                    "403": {
                        "description": "Forbidden - incorrect password"
                    },
                    "409": {
                        "description": "Conflict - 2FA already enabled"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/user/password": {
            "put": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Changes the password for the currently authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Change user password",
                "parameters": [
                    {
                        "description": "Password change request",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password changed successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request - validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid current password",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "security": [
                    {
                        "JWTBearerToken": []
                    }
                ],
                "description": "Returns a user by id with detailed channel membership information",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get user data by id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/controllers.UserResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "x-order": "0"
                },
                "username": {
                    "type": "string",
                    "x-order": "1"
                },
                "email": {
                    "type": "string",
                    "x-order": "2"
                },
                "max_logins": {
                    "type": "integer",
                    "x-order": "3"
                },
                "language_code": {
                    "type": "string",
                    "x-order": "4"
                },
                "language_name": {
                    "type": "string",
                    "x-order": "5"
                },
                "last_seen": {
                    "type": "integer",
                    "x-order": "6"
                },
                "totp_enabled": {
                    "type": "boolean",
                    "x-order": "7"
                },
                "channels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/controllers.ChannelMembership"
                    },
                    "x-order": "8"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWTBearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}